# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation My API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user or perusahaan mitra",
                "description": "Endpoint untuk melakukan login bagi pengguna dan perusahaan mitra.",
                "operationId": "login",
                "requestBody": {
                    "description": "Email dan password untuk login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model user"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user or perusahaan mitra",
                "description": "Endpoint untuk melakukan logout bagi pengguna dan perusahaan mitra.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Endpoint untuk melakukan pendaftaran pengguna baru.",
                "operationId": "register",
                "requestBody": {
                    "description": "Form pendaftaran",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model user"
                    }
                }
            }
        },
...
        "/api/jadwal/add": {
            "post": {
                "tags": [
                    "Jadwal"
                ],
                "summary": "Tambah data jadwal",
                "description": "Endpoint ini digunakan untuk menambah data jadwal.",
                "operationId": "tambahJadwal;",
                "requestBody": {
                    "description": "Form tambah jadwal",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "tanggal",
                                    "tempat",
                                    "user_ids"
                                ],
                                "properties": {
                                    "tanggal": {
                                        "type": "string"
                                    },
                                    "tempat": {
                                        "type": "string"
                                    },
                                    "user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jadwal berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Jadwal berhasil ditambahkan"
                                        },
                                        "jadwal": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/jadwal/{id}": {
            "get": {
                "tags": [
                    "Jadwal"
                ],
                "summary": "Get List jadwal",
                "description": "data jadwal",
                "operationId": "showBerita",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jadwal yang ingin diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return array model jadwal"
                    }
                }
            }
        },
        "/api/jadwal/update/{id}": {
            "post": {
                "tags": [
                    "Jadwal"
                ],
                "summary": "Update data jadwal",
                "description": "Update data jadwal berdasarkan ID",
                "operationId": "updateBerita",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jadwal yang ingin diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data yang akan diperbarui",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "tanggal": {
                                        "type": "string"
                                    },
                                    "tempat": {
                                        "type": "string"
                                    },
                                    "user_ids": {
                                        "type": "string",
                                        "format": "array"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model jadwal"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}